+++ symbolic-debuginfo/src/macho/mod.rs
@@ -13,6 +13,7 @@ use symbolic_common::{Arch, AsSelf, CodeId, DebugId, Uuid};
 
 use crate::base::*;
 use crate::dwarf::{Dwarf, DwarfDebugSession, DwarfError, DwarfSection, Endian};
+use crate::ParseObjectOptions;
 pub(crate) use mono_archive::{MonoArchive, MonoArchiveObjects};
 
 mod bcsymbolmap;
@@ -385,7 +386,7 @@ impl<'d> Parse<'d> for MachObject<'d> {
         Self::test(data)
     }
 
-    fn parse(data: &'d [u8]) -> Result<Self, MachError> {
+    fn parse_with_opts(data: &'d [u8], _opts: ParseObjectOptions) -> Result<Self, Self::Error> {
         Self::parse(data)
     }
 }
@@ -562,6 +563,7 @@ pub struct FatMachObjectIterator<'d, 'a> {
     iter: mach::FatArchIterator<'a>,
     remaining: usize,
     data: &'d [u8],
+    opts: ParseObjectOptions,
 }
 
 impl<'d> Iterator for FatMachObjectIterator<'d, '_> {
@@ -577,7 +579,10 @@ impl<'d> Iterator for FatMachObjectIterator<'d, '_> {
             Some(Ok(arch)) => {
                 let start = (arch.offset as usize).min(self.data.len());
                 let end = (arch.offset as usize + arch.size as usize).min(self.data.len());
-                Some(MachObject::parse(&self.data[start..end]))
+                Some(MachObject::parse_with_opts(
+                    &self.data[start..end],
+                    self.opts,
+                ))
             }
             Some(Err(error)) => Some(Err(MachError::new(error))),
             None => None,
@@ -593,11 +598,10 @@ impl std::iter::FusedIterator for FatMachObjectIterator<'_, '_> {}
 impl ExactSizeIterator for FatMachObjectIterator<'_, '_> {}
 
 /// A fat MachO container that hosts one or more [`MachObject`]s.
-///
-/// [`MachObject`]: struct.MachObject.html
 pub struct FatMachO<'d> {
     fat: mach::MultiArch<'d>,
     data: &'d [u8],
+    opts: ParseObjectOptions,
 }
 
 impl<'d> FatMachO<'d> {
@@ -608,8 +612,13 @@ impl<'d> FatMachO<'d> {
 
     /// Tries to parse a fat MachO container from the given slice.
     pub fn parse(data: &'d [u8]) -> Result<Self, MachError> {
+        Self::parse_with_opts(data, Default::default())
+    }
+
+    /// Tries to parse a fat MachO container from the given slice.
+    pub fn parse_with_opts(data: &'d [u8], opts: ParseObjectOptions) -> Result<Self, MachError> {
         mach::MultiArch::new(data)
-            .map(|fat| FatMachO { fat, data })
+            .map(|fat| FatMachO { fat, data, opts })
             .map_err(MachError::new)
     }
 
@@ -619,6 +628,7 @@ impl<'d> FatMachO<'d> {
             iter: self.fat.iter_arches(),
             remaining: self.fat.narches,
             data: self.data,
+            opts: self.opts,
         }
     }
 
@@ -639,7 +649,7 @@ impl<'d> FatMachO<'d> {
 
         let start = (arch.offset as usize).min(self.data.len());
         let end = (arch.offset as usize + arch.size as usize).min(self.data.len());
-        MachObject::parse(&self.data[start..end]).map(Some)
+        MachObject::parse_with_opts(&self.data[start..end], self.opts).map(Some)
     }
 }
 
@@ -749,12 +759,17 @@ impl<'d> MachArchive<'d> {
         }
     }
 
-    /// Tries to parse a Mach archive from the given slice.
+    /// Tries to parse a Mach archive from the given slice, with default options.
     pub fn parse(data: &'d [u8]) -> Result<Self, MachError> {
+        Self::parse_with_opts(data, Default::default())
+    }
+
+    /// Tries to parse a Mach archive from the given slice.
+    pub fn parse_with_opts(data: &'d [u8], opts: ParseObjectOptions) -> Result<Self, MachError> {
         Ok(Self(match Self::is_fat(data) {
-            Some(true) => MachArchiveInner::Archive(FatMachO::parse(data)?),
+            Some(true) => MachArchiveInner::Archive(FatMachO::parse_with_opts(data, opts)?),
             // Fall back to mach parsing to receive a meaningful error message from goblin
-            _ => MachArchiveInner::Single(MonoArchive::new(data)),
+            _ => MachArchiveInner::Single(MonoArchive::new(data, opts)),
         }))
     }